Shim open64() on solaris
This is a copy of the open() shim.  It's needed, though, since it
sets the is_regular[] to true for the opened fd, which is used by
the read() and write() shims.  sqlite3 was using an open64() call
followed by read() and write() so it was not using the correct
code path.