flow exercise application
It takes a flow XML as argument and uses ./xml subdirectory to load
plugins (if its there).  The flow is loaded and run in a dummy mode
with functions which operate the nodes and report on the guards.
Currently all guards are de-keyed (so multiple acquisitions will lockup
actually) which simplifies the interaction a bit.  The goal of this
tool is to provide a way to run the flow in an artificial way that
does not involve application code.  Some meaningful measurement and
runtime debugging may be possible with this.