ability for user to inject events
User programs often want to be able to inject events
onto the runtime queue from deep within their logic.
The typical way that this is done with an eventMux (fd and select)
mechanism that tickles a source.  This is wasteful since it
dedicates a thread to serving the work purpose.

This change uses the existing runtime::submitEvents() functionality
which was added for libdoor support.